refactor: extract dispatcher config reader and fine-tune base-model validator
